requirements.txt
boto3==1.16.53
botocore==1.19.53
Django==3.1.4
django-crispy-forms==1.10.0
django-storages==1.11.1
djangorestframework==3.12.2
Pillow==8.0.1
pylint==2.6.0
requests==2.25.1
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
urllib3==1.26.2
psycopg2-binary==2.8.5
